Move the validateHeaderName to the public HeaderNames object
Ths method is also useful for http4s.
